requirements.txt
aiohttp>=3.3.2
discord.py>=1.3.4
websockets>=8.1
python-chess>=0.31.3
